move exact tests from legacy
