+++ test/legacy/steadystate/test_steadystate.py
@@ -31,7 +31,7 @@ def _setup_system():
         )
         j_ops.append(nk.operator.LocalOperator(hi, sigmam, [i]))
 
-    #  Create the liouvillian
+    # Create the liouvillian
     lind = nk.operator.LocalLiouvillian(ha, j_ops)
     return hi, lind
 
@@ -70,37 +70,6 @@ def _setup_obs():
 
     obs = {"SigmaX": obs_sx}
     return obs
-
-
-def test_exact_ss_ed():
-    _, lind = _setup_system()
-
-    dm_ss = nk.exact.steady_state(lind, method="ed", sparse=True)
-    Lop = lind.to_sparse()
-
-    mat = np.abs(Lop @ dm_ss.reshape(-1))
-    print(mat)
-    assert np.all(mat == approx(0.0, rel=1e-4, abs=1e-4))
-    assert dm_ss.trace() - 1 == approx(0.0, rel=1e-5, abs=1e-8)
-
-    # dm_ss_d = nk.exact.steady_state(lind, method="ed", sparse=False)
-    # Lop = lind.to_sparse()
-
-    # mat = np.abs(dm_ss - dm_ss_d)
-    # print(mat)
-    # assert np.all(mat == approx(0.0, rel=1e-4, abs=1e-4))
-
-
-def test_exact_ss_iterative():
-    _, lind = _setup_system()
-
-    dm_ss = nk.exact.steady_state(lind, sparse=True, method="iterative", tol=1e-5)
-    Lop = lind.to_linear_operator()
-
-    mat = np.abs(Lop @ dm_ss.reshape(-1))
-    print(mat)
-    assert np.all(mat == approx(0.0, rel=1e-5, abs=1e-5))
-    assert dm_ss.trace() - 1 == approx(0.0, rel=1e-5, abs=1e-5)
 
 
 def test_ss_advance():
